Рефакторинг/декомпозиция IndexController выделение отдельного слоя сервисов.
